Move following filter to public
The filter in rollup is overriding filters passed into rollup elsewhere
in the code, so instead move it to public and pass as a parameter. Due
to fact that refresh is needed to update view when the filter changes,
using new parameter observableFilter that hooks up the callback for
refresh.