Add User Font Size Setting
Adds a selection for the user to change the font size from the settings
panel. Persists in localstorage and updates on changes as language and
theme. Selection range: 8,10,12,14,16,18,20 px.

Most locales updated except `ki.json` which is in english.

Font size is set in a second `style` element, appended immediately after
the theme is applied in `main-window.js`, overwriting the theme font
size. I considered finding a more clever way to replace the original
font size declaration in place but haven't got very far yet.