Redact private key from startup output
Previously we did this output *before* adding the keys to the config,
but recently there were changes made that required the keys to be added
to the config before the config was finished. This meant that the config
being output contained private keys, which seems like a major security
hazard for debugging.

This commit changes the code so that we make a copy of the full config,
set `redactedConfig.keys.private = null`, and then output the redacted
configuration safely without any secrets exposed.