Fix pull-resume pagination bug
The problem seems to be caused by the marker counter getting ahead of
the limit constrained by `pull.take()`. This moves both the filter and
the `pull.take()` before the marker counter, which ensures that if the
counters become out of sync the marker counter can only be behind --
never ahead.