refactor: move files out of project root
The intention here is to make the project root a bit less busy and
stimulating so that new users can get a better feel for what's going on
in the project. This isn't perfect, but I think it's an improvement and
it reduces the number of files that the user is greeted with when
looking at the Patchwork repo.