refactor: restyle to use reduced line length
This restyles Patchwork to use a much smaller line length which is in
line with best practices regarding typography. This also includes a
change to the layout of the profile page so that elements have more room
to breathe as well as a change to the font-size and the margins of <h1>
through <h6> so that they don't have such a stark contrast with the rest
of the text rendered.

Resolves https://github.com/ssbc/patchwork/issues/1102