feat: make alt text visually accessible
I've seen that Dan has been adding captions to lots of his images
lately, which is super rad, but it's a bummer that he has to put the
descriptions in both the alt text *and* a secondary paragraph to make it
visually accessible.

This change takes the alt text and puts it into a `<figcaption>` that's
hidden from accessibility tools, that way screen readers see the image
alt text but people using Scuttlebutt visually are still able to see the
alt text that the community is adding to their images.

My hope is that this highlights the utility behind alt text and inspires
people to describe their visual data in a way that can be better
understood by accessibility tools.