Change to throttle sidebar UI on public page
When connections change quickly the sidebar can flash and strobe, which
is annoying at best and inaccessible at worst. This commit throttles the
UI changes at a maximum of 1 change per second, which reduces the
unwanted behavior.

Note: The `throttle()` method from Mutant doesn't have tests or
documentation, so this is my best guess at how it's meant to be used
from reading the source code.