Remove remove-markdown dependency.
We already had a method to make a short version of a markdown text.
It was stuck deep in a depject module, so I ripped it out of there and put
it into its own module.
It is now require()d in the depject module and in the new bio preview code.