Public web hosting support
It’s a profile setting, because it’s an “about” message. I’m not sure that’s the most logical place to put it (vs settings).

This PR does not yet work, it isn’t able to detect the current state of the publicWebHosting value and I don’t know why. I’d love feedback on this.