Bump electron.
To the latest version.
Seems to work, mostly. There's some slowness happening in the search
suggestions & completions, and of course there's still no context menu.
But posting & scuttling & reading still works, so there's that.

This still relies on the `remote` module, which is deprecated and considered
harmful. Need to work around that in the future, which... makes sense.