Port context menu to electron 11 & electron-context-menu
The old menu was handling more than we needed: it was doing spell-checking,
which has been built into electron since v8.0.0, and it was also overly
complicated. The new version should be a bit easier to read.
Avoiding use of the remote module means that a lot of things have to be
handled via explicit IPC (instead of relying on implicit proxy objects from
the remote module) and especially for the "Copy Message Text" and "Copy
Message Reference" items and such I had to hack a bit. As a result, this
*might* still have race conditions that would mean that text doesn't get
copied, or an outdated version of it.