Add mnemonic key export.
This comes with a quiz to make it hard for people to misunderstand what
it does and doesn't do.
The fear was/is that providing this will result in a wave of people
with forked feeds landing at our/other projects' gates, asking for help
with an unfixable problem.
So the text is quite long, and the quiz spot-checks that it was actually
read and understood. It also makes folks type out actively (no
copy-paste!) a few sentences that should make it very hard to ignore the
risks involved.
It's only in english so far...